requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.1.3
requests==2.32.3
pandas==2.2.3
matplotlib==3.9.2
python-dotenv==1.0.1
email_validator==2.2.0
openpyxl==3.1.5
pytest==8.3.4
pytest-cov==6.0.0
